refactor: update remaining planter references
