+++ src/models/auth.js
@@ -10,8 +10,8 @@ const POLICIES = {
   MANAGER_USER: 'manager_user',
   LIST_TREE: 'list_tree',
   APPROVE_TREE: 'approve_tree',
-  LIST_GROWER: 'list_GROWER',
-  MANAGE_PLANTER: 'manage_planter',
+  LIST_GROWER: 'list_planter',
+  MANAGE_GROWER: 'manage_planter',
 };
 
 function hasPermission(user, p) {
